+++ app/src/main/java/com/example/munkkistracker/Munkki.java
@@ -2,11 +2,11 @@ package com.example.munkkistracker;
 
 public interface Munkki {  //rajapinta munkeille
 
-    public float getRasva();
+    public double getRasva();
 
     public int getKcal();
 
-    public float getSokeri();
+    public double getSokeri();
